src/startup_service/requirements.txt
# PyGent Factory System Startup Service Dependencies
# Production-ready FastAPI application with async support

# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0
alembic==1.12.1

# Authentication and Security
pyjwt[crypto]==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.7

# HTTP Client and WebSocket
aiohttp==3.9.1
websockets==12.0
httpx==0.25.2

# Configuration and Environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Monitoring and Metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
structlog==23.2.0

# Task Queue and Background Jobs
celery[redis]==5.3.4
redis==5.0.1

# Docker and Container Management
docker==6.1.3
docker-compose==1.29.2

# Utilities and Helpers
click==8.1.7
rich==13.7.0
typer==0.9.0
jinja2==3.1.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Production WSGI/ASGI
gunicorn==21.2.0

# System Monitoring
psutil==5.9.6

# Date and Time
python-dateutil==2.8.2

# JSON and Data Processing
orjson==3.9.10
msgpack==1.0.7

# Async Extensions
asyncio-mqtt==0.16.1
aiofiles==23.2.1
aioredis==2.0.1

# Validation and Serialization
marshmallow==3.20.1
cerberus==1.3.5

# Network and HTTP
requests==2.31.0
urllib3==2.1.0

# Logging and Observability
loguru==0.7.2
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# File System and Path Handling
pathlib2==2.3.7
watchdog==3.0.0

# Concurrency and Threading
asyncio-throttle==1.0.2
aiolimiter==1.1.0

# Error Handling and Retry
tenacity==8.2.3
backoff==2.2.1

# Environment and System
python-decouple==3.8
environs==10.0.0
